Slice 18: Reorganize shared components by intent
- Move flat components/shared/ + loose components/EntityLink and
  components/MobilePreview into intent-grouped subfolders:
    components/shared/feedback/  -> ConfirmDialog
    components/shared/data/      -> CSVExport, EntityLink, StatCard
    components/shared/preview/   -> MobilePreview
- Add file-header docblocks + JSDoc to every shared component while
  moving them.
- Add a tiny barrel index.js to each subfolder plus a top-level
  components/shared/index.js that re-exports everything.
- Update all 18 import sites to pull from the new intent barrels
  (e.g. import { ConfirmDialog } from "../components/shared/feedback").
- Verified npm run build succeeds with the new layout.
