+++ src/admin-app/src/components/shared/data/index.js
@@ -1,0 +1,8 @@
+/**
+ * @file index.js
+ * @summary Barrel for shared data-display components (tables, cards, exports).
+ */
+
+export { default as CSVExport } from "./CSVExport";
+export { default as EntityLink } from "./EntityLink";
+export { default as StatCard } from "./StatCard";
