+++ src/admin-app/src/components/shared/feedback/ConfirmDialog.jsx
@@ -1,12 +1,29 @@
+/**
+ * @file ConfirmDialog.jsx
+ * @summary Reusable destructive-confirm modal.
+ *
+ * The "Confirm" button always renders in the error color so it reads as
+ * destructive at a glance. Caller supplies the copy via `title` /
+ * `message` so the dialog stays domain-agnostic.
+ */
+
 import {
+  Button,
   Dialog,
-  DialogTitle,
+  DialogActions,
   DialogContent,
   DialogContentText,
-  DialogActions,
-  Button,
+  DialogTitle,
 } from "@mui/material";
 
+/**
+ * @param {object} props
+ * @param {boolean} props.open
+ * @param {string} props.title
+ * @param {string} props.message - Body copy. Plain string (no JSX).
+ * @param {() => void} props.onConfirm
+ * @param {() => void} props.onCancel
+ */
 export default function ConfirmDialog({ open, title, message, onConfirm, onCancel }) {
   return (
     <Dialog open={open} onClose={onCancel} aria-describedby="confirm-dialog-description">
