+++ src/admin-app/src/components/shared/feedback/index.js
@@ -1,0 +1,6 @@
+/**
+ * @file index.js
+ * @summary Barrel for shared feedback components (modals, alerts, toasts).
+ */
+
+export { default as ConfirmDialog } from "./ConfirmDialog";
