+++ src/admin-app/src/components/shared/preview/index.js
@@ -1,0 +1,6 @@
+/**
+ * @file index.js
+ * @summary Barrel for preview/visualization components (mobile/email/etc.).
+ */
+
+export { default as MobilePreview } from "./MobilePreview";
