+++ src/admin-app/src/features/dashboard/widgets/RecentActivityWidget.jsx
@@ -1,6 +1,6 @@
 import { Box, Typography, Chip, Stack } from "@mui/material";
 
-import EntityLink from "../../../components/EntityLink";
+import { EntityLink } from "../../../components/shared/data";
 
 export default function RecentActivityWidget({ data }) {
   if (!data.recentActivity?.length) {
