+++ src/admin-app/src/features/reports/ReportsPage.jsx
@@ -35,8 +35,7 @@ import {
   Alert,
 } from "@mui/material";
 
-import EntityLink from "../../components/EntityLink";
-import CSVExport from "../../components/shared/CSVExport";
+import { CSVExport, EntityLink } from "../../components/shared/data";
 import { getChallenges, getUsers, getParticipation, getActions } from "../../data/api";
 
 export default function ReportsPage() {
